fix: fix bug with statistics (views count)

+++ mainapp/src/main/java/ru/practicum/ewm/event/EventShortDto.java
@@ -32,6 +32,6 @@ public class EventShortDto {
     @NotNull
     private String title;
     private Integer views;
-    private Float rating;
+    private double rating;
 
 }
